Extract enum traits to a separate header
